utils for date handling, removed disable create date
