+++ client/src/components/Forms/DailyProgressAndActivity.js
@@ -13,6 +13,7 @@ import { FormSavedAlert } from "../../utils/FormSavedAlert";
 import { isAdminUser } from "../../utils/AdminReportingRoles";
 import TextareaAutosize from "react-textarea-autosize";
 import { Container, Row, Col } from "react-bootstrap";
+import { HandleFieldInputDate } from "../../utils/HandleFieldInputDate";
 /*
   missing from form
     "Restricted field Trip"
@@ -106,20 +107,6 @@ class DailyProgressAndActivity extends Component {
     this.setState(stateObj);
   };
 
-  handleFieldInputDate = (event) => {
-    var stateObj = {};
-    if (event.target.id.indexOf(".") > -1) {
-      let level1Obj = event.target.id.split(".")[0];
-      let level2Obj = event.target.id.split(".")[1];
-
-      let nestedProperty = { ...this.state[level1Obj] };
-      nestedProperty[level2Obj] = event.target.value;
-      stateObj[level1Obj] = nestedProperty;
-    } else {
-      stateObj[event.target.id] = event.target.value.concat(':00.000Z');
-    }
-    this.setState(stateObj);
-  };
 
   resetForm = () => {
     this.setState({
@@ -370,6 +357,13 @@ class DailyProgressAndActivity extends Component {
     }
   };
 
+  setRootState = (body) => {
+    const stateCopy = { ...this.state, ...body };
+    this.setState({
+      ...stateCopy,
+    });
+  };
+
   render() {
     if (!this.props.valuesSet) {
       return (
@@ -402,8 +396,8 @@ class DailyProgressAndActivity extends Component {
                   {`${new Date(this.state.lastEditDate)
                     .toTimeString()
                     .replace(/\s.*/, "")} - ${new Date(
-                    this.state.lastEditDate
-                  ).toDateString()}`}
+                      this.state.lastEditDate
+                    ).toDateString()}`}
                 </i>
               ) : (
                 "-"
@@ -425,18 +419,10 @@ class DailyProgressAndActivity extends Component {
           ) : (
 
             <Container className="print-container">
-              <div className="form-group logInInputField">
-                <label className="control-label hide hide-on-print">
-                  Create Date
-                </label>{" "}
-                <input
-                  onChange={this.handleFieldInputDate}
-                  id="createDate"
-                  value={this.state.createDate.slice(0, -8)}
-                  className="form-control"
-                  type="datetime-local"
-                />{" "}
-              </div> 
+              <HandleFieldInputDate
+                setRootState={this.setRootState}
+                rootState={this.state}
+              />
               <div className="form-group logInInputField">
                 {" "}
                 <label className="control-label">Child's Name</label>{" "}
@@ -476,7 +462,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.personal_hygiene}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -488,7 +474,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.dressing}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -500,7 +486,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.table_mannders}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -514,7 +500,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.clothes_maintenace}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -526,7 +512,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.self_feeding}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -540,7 +526,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.care_of_property}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -554,7 +540,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.maintenace_of_personal_space}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -568,7 +554,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.household_chorse}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                 </Col>
@@ -591,7 +577,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.informal_counseling}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -605,7 +591,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.verbal_redirection}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -617,7 +603,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.modeling}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -631,7 +617,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.supervised_separation}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -645,7 +631,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.provider_feedback_to_client}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -659,7 +645,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.positive_reinforcement}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -673,7 +659,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.other}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group input-header">
@@ -692,7 +678,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.home_restrictions}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -706,7 +692,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.restricted_leisure_activity}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -718,7 +704,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.no_allowance}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -732,7 +718,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.other2}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                 </Col>
@@ -751,7 +737,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.no_of_home_incidents}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -765,7 +751,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.no_of_home_serious_incidents}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -779,7 +765,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.no_of_home_restraints}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -793,7 +779,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.no_of_school_incidents}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -807,7 +793,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.no_of_school_restraints}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -821,7 +807,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.illness_injury}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -835,7 +821,7 @@ class DailyProgressAndActivity extends Component {
                       value={this.state.level_of_supervison}
                       className="form-control"
                       type="text"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     />{" "}
                   </div>
                   <div className="form-group logInInputField">
@@ -848,7 +834,7 @@ class DailyProgressAndActivity extends Component {
                       id="summary_of_daily_schedule"
                       value={this.state.summary_of_daily_schedule}
                       className="form-control"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     ></TextareaAutosize>
                   </div>
                   <div className="form-group logInInputField">
@@ -861,7 +847,7 @@ class DailyProgressAndActivity extends Component {
                       id="summary_of_behavior_at_school"
                       value={this.state.summary_of_behavior_at_school}
                       className="form-control"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     ></TextareaAutosize>
                   </div>
                   <div className="form-group logInInputField">
@@ -874,7 +860,7 @@ class DailyProgressAndActivity extends Component {
                       id="summary_of_behavior_at_home"
                       value={this.state.summary_of_behavior_at_home}
                       className="form-control"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     ></TextareaAutosize>
                   </div>
                   <div className="form-group logInInputField">
@@ -887,7 +873,7 @@ class DailyProgressAndActivity extends Component {
                       id="therapeutic_recreational"
                       value={this.state.therapeutic_recreational}
                       className="form-control"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     ></TextareaAutosize>
                   </div>
                   <div className="form-group logInInputField">
@@ -900,7 +886,7 @@ class DailyProgressAndActivity extends Component {
                       id="therapeutic_value"
                       value={this.state.therapeutic_value}
                       className="form-control"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     ></TextareaAutosize>
                   </div>
                   <div className="form-group logInInputField">
@@ -913,7 +899,7 @@ class DailyProgressAndActivity extends Component {
                       id="phone_calls_or_visits"
                       value={this.state.phone_calls_or_visits}
                       className="form-control"
-                      disabled = {this.state.childSelected ? false : true}
+                      disabled={this.state.childSelected ? false : true}
                     ></TextareaAutosize>
                   </div>
                   <FormError errorId={this.props.id + "-error"} />
@@ -923,36 +909,36 @@ class DailyProgressAndActivity extends Component {
                   ></div>
                 </Col>
               </Row>
-              
+
               <Row className="save-submit-row">
-                <div style={{display:"flex", width:"46%"}}>
+                <div style={{ display: "flex", width: "46%" }}>
                   <button
-                      className="lightBtn hide hide-on-print save-submit-btn"
-                      style={{width:"100%"}}
-                      disabled={this.state.childSelected ? false : true}
-                      onClick={() => {
-                        this.validateForm(true);
-                      }}
-                    >
-                      Finish Later
-                    </button>
+                    className="lightBtn hide hide-on-print save-submit-btn"
+                    style={{ width: "100%" }}
+                    disabled={this.state.childSelected ? false : true}
+                    onClick={() => {
+                      this.validateForm(true);
+                    }}
+                  >
+                    Finish Later
+                  </button>
                 </div>
-                <div style={{display:"flex", width:"46%"}}>
+                <div style={{ display: "flex", width: "46%" }}>
                   <button
-                      className="darkBtn hide hide-on-print save-submit-btn"
-                      style={{width:"100%"}}
-                      disabled={this.state.childSelected ? false : true}
-                      onClick={() => {
-                        this.validateForm(false);
-                      }}
-                    >
-                      Submit
+                    className="darkBtn hide hide-on-print save-submit-btn"
+                    style={{ width: "100%" }}
+                    disabled={this.state.childSelected ? false : true}
+                    onClick={() => {
+                      this.validateForm(false);
+                    }}
+                  >
+                    Submit
                   </button>
                 </div>
 
-                 
+
               </Row>
-              
+
             </Container>
           )}
         </div>
@@ -1004,7 +990,6 @@ class DailyProgressAndActivity extends Component {
                     value={this.state.createDate.slice(0, -8)}
                     className="form-control hide-on-print"
                     type="datetime-local"
-                    disabled
                   />{" "}
                 </div>
                 <div className="form-group logInInputField">
@@ -1327,7 +1312,7 @@ class DailyProgressAndActivity extends Component {
                         id="no_of_home_restraints"
                         className="form-control"
                         type="text"
-                        
+
                       />{" "}
                     </div>
                     <div className="form-group logInInputField">
@@ -1485,9 +1470,9 @@ class DailyProgressAndActivity extends Component {
                 </Row>
               </Container>
             )}
-            
+
             <div className="sigSection"
-              style= {{ display: this.state.status === 'IN PROGRESS' ? 'none' : 'block' }}
+              style={{ display: this.state.status === 'IN PROGRESS' ? 'none' : 'block' }}
             >
               <label className="control-label">Signature</label>{" "}
               <div
@@ -1518,10 +1503,10 @@ class DailyProgressAndActivity extends Component {
               <>
                 <FormError errorId={this.props.id + "-error"} />
                 <Row className="save-submit-row">
-                    <div style={{ display: "flex", width: "46%" }}>
+                  <div style={{ display: "flex", width: "46%" }}>
                     <button
                       className="lightBtn hide hide-on-print save-submit-btn"
-                      style={{ 
+                      style={{
                         width: "100%",
                         display: this.state.status === 'COMPLETED' ? "none" : "block",
                       }}
@@ -1533,18 +1518,18 @@ class DailyProgressAndActivity extends Component {
                     </button>
                   </div>
 
-                <div style={{ display: "flex", width: "46%" }}>
-                  <button
-                    className="darkBtn hide hide-on-print save-submit-btn"
-                    style={{ width: "100%" }}
-                    onClick={() => {
-                      this.validateForm(false);
-                    }}
-                  >
-                    Submit
-                  </button>
-                </div>
-              </Row>
+                  <div style={{ display: "flex", width: "46%" }}>
+                    <button
+                      className="darkBtn hide hide-on-print save-submit-btn"
+                      style={{ width: "100%" }}
+                      onClick={() => {
+                        this.validateForm(false);
+                      }}
+                    >
+                      Submit
+                    </button>
+                  </div>
+                </Row>
               </>
             )}
           </div>
