+++ client/src/components/Forms/IncidentReport.js
@@ -14,6 +14,7 @@ import { isAdminUser } from "../../utils/AdminReportingRoles";
 import TextareaAutosize from "react-textarea-autosize";
 import StaffOption from "../../utils/StaffOption.util";
 import { Container, Col, Row } from "react-bootstrap";
+import { HandleFieldInputDate } from "../../utils/HandleFieldInputDate";
 var interval = 0; // used for autosaving
 let initAutoSave = false;
 class IncidentReport extends Component {
@@ -93,21 +94,6 @@ class IncidentReport extends Component {
       stateObj[level1Obj] = nestedProperty;
     } else {
       stateObj[event.target.id] = event.target.value;
-    }
-    this.setState(stateObj);
-  };
-
-  handleFieldInputDate = (event) => {
-    var stateObj = {};
-    if (event.target.id.indexOf(".") > -1) {
-      let level1Obj = event.target.id.split(".")[0];
-      let level2Obj = event.target.id.split(".")[1];
-
-      let nestedProperty = { ...this.state[level1Obj] };
-      nestedProperty[level2Obj] = event.target.value;
-      stateObj[level1Obj] = nestedProperty;
-    } else {
-      stateObj[event.target.id] = event.target.value.concat(':00.000Z');
     }
     this.setState(stateObj);
   };
@@ -401,6 +387,13 @@ class IncidentReport extends Component {
     }
   };
 
+  setRootState = (body) => {
+    const stateCopy = { ...this.state, ...body };
+    this.setState({
+      ...stateCopy,
+    });
+  };
+
   render() {
     if (!this.props.valuesSet) {
       return (
@@ -457,18 +450,10 @@ class IncidentReport extends Component {
             <Container className="print-container">
               <Row>
                 <Col xs={12} className="print-column">
-                  <div className="form-group logInInputField">
-                    <label className="control-label hide-on-print">
-                      Create Date
-                    </label>{" "}
-                    <input
-                      onChange={this.handleFieldInputDate}
-                      id="createDate"
-                      value={this.state.createDate.slice(0, -8)}
-                      className="form-control hide-on-print"
-                      type="datetime-local"
-                    />{" "}
-                  </div>
+                  <HandleFieldInputDate
+                    setRootState={this.setRootState}
+                    rootState={this.state}
+                  />
                 </Col>
               </Row>
 
@@ -1035,7 +1020,6 @@ class IncidentReport extends Component {
                         value={this.state.createDate.slice(0, -8)}
                         className="form-control hide-on-print"
                         type="datetime-local"
-                        disabled
                       />{" "}
                     </div>
                   </Col>
