+++ client/src/components/Forms/TreatmentPlan72.js
@@ -13,6 +13,7 @@ import { FormSavedAlert } from "../../utils/FormSavedAlert";
 import { isAdminUser } from "../../utils/AdminReportingRoles";
 import TextareaAutosize from "react-textarea-autosize";
 import { Row } from "react-bootstrap";
+import { HandleFieldInputDate } from "../../utils/HandleFieldInputDate";
 
 var interval = 0; // used for autosaving
 let initAutoSave = false;
@@ -209,22 +210,6 @@ class TreatmentPlan72 extends Component {
     }
     this.setState(stateObj);
   };
-
-  handleFieldInputDate = (event) => {
-    var stateObj = {};
-    if (event.target.id.indexOf(".") > -1) {
-      let level1Obj = event.target.id.split(".")[0];
-      let level2Obj = event.target.id.split(".")[1];
-
-      let nestedProperty = { ...this.state[level1Obj] };
-      nestedProperty[level2Obj] = event.target.value;
-      stateObj[level1Obj] = nestedProperty;
-    } else {
-      stateObj[event.target.id] = event.target.value.concat(':00.000Z');
-    }
-    this.setState(stateObj);
-  };
-
 
   resetForm = () => {
     this.setState({
@@ -713,6 +698,13 @@ class TreatmentPlan72 extends Component {
     }
   };
 
+  setRootState = (body) => {
+    const stateCopy = { ...this.state, ...body };
+    this.setState({
+      ...stateCopy,
+    });
+  };
+
   render() {
     if (!this.props.valuesSet) {
       return (
@@ -769,16 +761,10 @@ class TreatmentPlan72 extends Component {
               </div>
             ) : (
               <div>
-                <div className="form-group logInInputField">
-                  <label className="control-label">Create Date</label>{" "}
-                  <input
-                    onChange={this.handleFieldInputDate}
-                    id="createDate"
-                    value={this.state.createDate.slice(0, -8)}
-                    className="form-control"
-                    type="datetime-local"
-                  />{" "}
-                </div>
+                <HandleFieldInputDate
+                  setRootState={this.setRootState}
+                  rootState={this.state}
+                />
                 <div className="form-group logInInputField">
                   {" "}
                   <label className="control-label">Child's Name</label>{" "}
